# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Cricket Box API
  version: 1.0.0
  description: |
    ## Cricket Box - Digital Box Cricket Management Platform
    
    Complete API documentation for Cricket Box backend system.
    
    ### Features
    - User Authentication (Email/Password & OTP)
    - Room Management
    - Team Setup
    - Live Match Scoring
    - Statistics & Leaderboards
    - Admin Dashboard
    
  contact:
    name: Cricket Box Support
    email: support@cricketbox.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.cricketbox.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT token
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: objectId
        username:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        avatar:
          type: string
          format: uri
        phone:
          type: string
        role:
          type: string
          enum: [admin, host, player, umpire, viewer]
        status:
          type: string
          enum: [active, inactive, blocked, pending_verification]
        isEmailVerified:
          type: boolean
        statistics:
          $ref: '#/components/schemas/UserStatistics'
        createdAt:
          type: string
          format: date-time
    UserStatistics:
      type: object
      properties:
        matchesPlayed:
          type: integer
          default: 0
        matchesWon:
          type: integer
          default: 0
        matchesLost:
          type: integer
          default: 0
        totalRuns:
          type: integer
          default: 0
        totalBallsFaced:
          type: integer
          default: 0
        highestScore:
          type: integer
          default: 0
        centuries:
          type: integer
          default: 0
        fifties:
          type: integer
          default: 0
        fours:
          type: integer
          default: 0
        sixes:
          type: integer
          default: 0
        notOuts:
          type: integer
          default: 0
        totalWickets:
          type: integer
          default: 0
        totalOversBowled:
          type: number
          default: 0
        totalRunsConceded:
          type: integer
          default: 0
        catches:
          type: integer
          default: 0
        runOuts:
          type: integer
          default: 0
        stumpings:
          type: integer
          default: 0

    # Auth Schemas
...
    Room:
      type: object
      properties:
        id:
          type: string
          format: objectId
        name:
          type: string
        code:
          type: string
          minLength: 6
          maxLength: 6
        description:
          type: string
        host:
          $ref: '#/components/schemas/User'
        status:
          type: string
          enum: [waiting, team_setup, ready, in_match, completed, closed]
        settings:
          $ref: '#/components/schemas/RoomSettings'
        participants:
          type: array
          items:
            $ref: '#/components/schemas/Participant'
        guestParticipants:
          type: array
          items:
            $ref: '#/components/schemas/GuestParticipant'
        teamA:
          $ref: '#/components/schemas/Team'
        teamB:
          $ref: '#/components/schemas/Team'
        umpire:
          $ref: '#/components/schemas/Umpire'
        currentMatch:
          type: string
          format: objectId
        createdAt:
          type: string
          format: date-time
    RoomSettings:
      type: object
      properties:
        overs:
          type: integer
          minimum: 1
          maximum: 50
          default: 6
        playersPerTeam:
          type: integer
          minimum: 2
          maximum: 11
          default: 6
        maxParticipants:
          type: integer
          minimum: 2
          maximum: 30
          default: 20
        isPrivate:
          type: boolean
          default: false
        allowGuests:
          type: boolean
          default: true
        wideRuns:
          type: integer
          default: 1
        noBallRuns:
          type: integer
          default: 1
        noBallFreehit:
          type: boolean
          default: true
    Participant:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        joinedAt:
          type: string
          format: date-time
        isReady:
          type: boolean
    GuestParticipant:
      type: object
      properties:
        name:
          type: string
        guestId:
          type: string
        joinedAt:
          type: string
          format: date-time
        isReady:
          type: boolean
    Team:
      type: object
      properties:
        name:
          type: string
        players:
          type: array
          items:
            $ref: '#/components/schemas/TeamPlayer'
    TeamPlayer:
      type: object
      properties:
        user:
          type: string
          format: objectId
        isGuest:
          type: boolean
        guestName:
          type: string
        guestId:
          type: string
        isCaptain:
          type: boolean
    Umpire:
      type: object
      properties:
        user:
          type: string
          format: objectId
        isGuest:
          type: boolean
        guestName:
          type: string
        guestId:
          type: string
    CreateRoomRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 50
          example: "Weekend Cricket"
        description:
          type: string
          maxLength: 500
          example: "Friendly match"
        settings:
          type: object
          properties:
            overs:
              type: integer
              example: 6
            playersPerTeam:
              type: integer
              example: 6
            isPrivate:
              type: boolean
              example: false
            password:
              type: string
              example: "secret123"
    JoinRoomRequest:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 6
          example: "ABC123"
        password:
          type: string
          example: "secret123"

    # Match Schemas
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errorCode:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
  responses:
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /rooms:
    post:
      tags: [Rooms]
      summary: Create a new room
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: Room created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          room:
                            $ref: '#/components/schemas/Room'

    get:
      tags: [Rooms]
      summary: Get all rooms
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: myRooms
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of rooms
  /rooms/join:
    post:
      tags: [Rooms]
      summary: Join a room by code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinRoomRequest'
      responses:
        '200':
          description: Joined room successfully
        '400':
          description: Invalid room code
        '409':
          description: Already in room
  /rooms/{roomId}:
    get:
      tags: [Rooms]
      summary: Get room by ID
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Room details
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags: [Rooms]
      summary: Close room (Host only)
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Room closed
        '403':
          description: Host only
  /rooms/{roomId}/teams/assign:
    post:
      tags: [Rooms]
      summary: Assign player to team
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [team]
              properties:
                team:
                  type: string
                  enum: [teamA, teamB]
                userId:
                  type: string
                guestId:
                  type: string
                isCaptain:
                  type: boolean
      responses:
        '200':
          description: Player assigned
  /rooms/{roomId}/match/start:
    post:
      tags: [Rooms]
      summary: Start a match
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Match started
        '400':
          description: Room not ready

  # Match Endpoints
...